overlay coreos-base/coreos-oem-gce: Pull in sys-apps/gentoo-functions too
Previously the sys-apps/gentoo-functions package was pulled in into
the GCE image by the sys-libs/glibc package. After the sys-libs/glibc
package update, the dependency disappeared. This resulted in
gentoo-functions not being installed and the build to fail a check for
dangling symlinks:

broken link: b'/etc/init.d/functions.sh' -> b'../../lib/gentoo/functions.sh'
ERROR   build_oem_aci: test_image_content: Failed symlink check

The `/etc/init.d/functions.sh` symlink is installed by
sys-apps/baselayout in postinst phase. The package also has a
dependency on sys-apps/gentoo-functions, but it is not pulled in into
the image, because baselayout is installed specially - without
dependencies.

It would probably be better to just drop the symlink, but it's here
for compatibility purposes. We also can't remove the symlink in the
manglefs script, as it gets executed after the filesystem checks.