net-misc/ena-driver: Allow installing as kernel builtin
This change installs the module sources and symlinks the module into the kernel
tree for building together in-tree kernel modules. This replaces the ena driver
in the kernel which is an older version (2.1.0) as far as reported version
goes. This is required for the module to be correctly signed for secure boot.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>