sys-kernel/bootengine: Change backport to avoid new stat specifiers
Newer versions of stat support %Hr and %Lr for major and minor device
types in decimal. Old stat only has %t and %T for the major and minor
device type in hex. This was causing the rule to break on Flatcar
Stable.
Change the backport to avoid the new stat decimal specifiers but instead
use the hex specifiers (Note: t is major and T in minor, but even when
one would confuse this, in our case both are 0).