overlay coreos/user-patches: Shuffle the SELinux patches
All the sec-policy/selinux-* packages contain policies from the same
tarball. Which means that for the sake of consistency we should be
applying our patches for every sec-policy/selinux- package. Currently
we have six such packages, so for each of those packages have a
symlink that points to the common selinux patches directory.