build_dev_binpkgs: Ensure adevcontainer binpkgs are built
This change introduces build_dev_binpkgs, a script to build binary
packages for all dependencies of the devcontainer. This works around an
issue with build_packages, which doesn't - leading to build issues with
the devcontainer later on. This particularly happens for more complex
builds with the devcontainer.

Additionally, a call to build_dev_binpkgs has been added to the package
publishing step in ci-automation before binary packages are published.

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>