overlay profiles: Mask installkernel flag in sys-apps/debianutils
This flag was masked for some time already in Gentoo profiles, and the
package that was supposed to be pulled was renamed in the meantime and
we didn't notice. Now the mask was moved to a different Gentoo
profile, so it stopped being masked in our case and the builds started
to fail, because there is no such package like
sys-kernel/installkernel in portage-stable (we have
sys-kernel/installkernel-gentoo).

Mask the flag for SDK. We will drop the installkernel package in
following commit.