changelog: remove cloudsigma custom cloudinit service configuration
CloudSigma uses a serial port to transfer metadata into a running
instance. Because the service was overwritten with the configuration
"restart" it was possible that the previous instance got interrupted.
This left the serial connection in an non-deterministic state as the
next user would get output from the previous request. The service in
general was not needed as the cloud oem is already set to cloudsigma.