nvidia-drivers: Improve match for NVIDIA GPUs
The logic for checking if we need to probe depends on the device showing up
with the product type in lspci output, which doesn't hold for all sorts of
GPUs. The NVIDIA_PRODUCT_TYPE is used for fetching the drivers and is "tesla"
across datacenter GPUs.

Switch to matching on vendor id and device class. The values are the same
ones that the nvidia driver binds to.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>