coreos-kernel.eclass: Install host arch Makefiles into source dir
Gentoo's linux-info.eclass tries to check the kernel version by
including /lib/modules/.../source/Makefile and printing $(VERSION).  It
unsets ARCH= before doing this (no clue why) which works with a full
source tree but not with our minimized one. To fix this we need to
archive arch/x86/Makefile also for arm64.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>