overlay coreos/user-patches: Update our SELinux policies patch
We were adding some modifications in init module that were referring
to the unconfined module. But init module is a part of
selinux-base-policy package, so it's a thing that's installed before
unconfined module. So loading base policies failed, because unconfined
module wasn't yet installed. Thus move the modification to the
unconfined module, so it becomes a modification in the unconfined
module that refers to the init module.