coreos-cargo.eclass: Introduce variable to skip inherit
Gentoo's cargo based ebuilds don't cross-compile well, so we need a way
to inject coreos-cargo.eclass into them. Otherwise we'd need to fork the
ebuilds into coreos-overlay and maintain them ourselves.

The way we can do this is by sourcing the eclass from cros hooks and
overriding src_unpack to call the eclass implementation. Inheriting an
eclass is not allowed from hooks. Since we can't call any of the
inherits a variable is introduced to skip them and EXPORT_FUNCTIONS.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>