sys-kernel/coreos-modules: Switch PTP support to modules
PTP_1588_CLOCK defaults to =y, make it =m explicitly. Switch
PTP_1588_CLOCK_KVM to module as well, so tath it doesn't load on non-KVM
platforms. Enable PTP_1588_CLOCK_VMW so that PTP timesync to the host is
available there as well.

We don't need PTP support to be compiled into the kernel, it is not
essential for system boot.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>