coreos-base/misc-files: Add AWS time sync service to default ntp.conf
The goal is to enable ntpd by default on AWS with the AWS time sync
service. Enabling ntpd.service is the easy part. Enabling the correct
ntp server is trickier. I would love to ship an AWS ntp.conf
specifically, but this might interfere with user modifications to either
ntpd.service or ntp.conf.

The safest way to implement this is to add the AWS time sync address
(which is a link-local address) to the default ntp.conf. This will work
on AWS where ntpd will be enabled by default, and if a user on another
platform enables ntpd they will simply have a non-responsive source in
their server list - which is not a problem for ntpd.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>