coreos-base/misc-files: Add AWS/GPC time sync service to default ntp.conf
The goal is to enable ntpd by default on AWS/GCP with using the cloud provided
ntp endpoint.  Enabling ntpd.service is the easy part. Enabling the correct ntp
server is trickier. I would love to ship an OEM specific ntp.conf, but this
might interfere with user modifications to either ntpd.service or ntp.conf.

The safest way to implement this is to add the AWS/GCP time sync address
(which are link-local address) to the default ntp.conf. This will work
on AWS/GCP where ntpd will be enabled by default, and if a user on another
platform enables ntpd they will simply have a non-responsive source in
their server list - which is not a problem for ntpd.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>