Move and rename repos, upgrade to Catalyst 4, support SDK on arm64
Sorry for doing all this in one giant commit, but it was hard to
separate it out.

We had no arm64 SDK, so some cross-compiling or emulation was most
likely going to be needed to produce one. Catalyst 4 adds support for
building with QEMU, so I looked into upgrading. This turned out to be
very much slower than emulating the amd64 SDK on arm64, where an arm64
build could then be mostly run without emulation. We can't stay on
Catalyst 3 forever though, so I continued with the upgrade.

Despite being slow, I have kept support for building with QEMU using
Catalyst since it requires little code and may be useful to somebody.

Catalyst 4 has totally changed the way repositories are handled. It only
works when the name of the directory containing the repository matches
the configured name of that repository. This was not the case for us,
with the coreos repository residing in the coreos-overlay directory. We
wanted to move and rename our repositories anyway, so they are now known
as gentoo-subset and flatcar-overlay, and they live under scripts/repos.
Using the same name as upstream Gentoo would have been problematic, and
just "flatcar" would have looked awkward in documentation.

Catalyst 4 also ingests the main repository snapshot as a squashfs
rather than a tarball. It features a utility to generate such a
snapshot, but it doesn't fit Flatcar well, particularly because it
expects each ebuild repository to reside at the top level of its own git
repository. It was very easy to call tar2sqfs manually though.

There were several places where we assumed that amd64 was native and
arm64 required emulation via QEMU. The scripts are now more
architecture-agnostic, paving the way for riscv support later.

We no longer set QEMU_LD_PREFIX because it prevents the SDK itself from
being emulated. It also assumes there is only one non-native target,
which won't be the case soon. bubblewrap does a better job of running
binaries under QEMU.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>