net-analyzer/netperf: use the MIT licensed version
As the net-analyzer/netperf stable version does not have a clear
License, the upstream Gentoo package has a masked ebuild that uses the
source from Github commit:
https://github.com/HewlettPackard/netperf/commit/3bc455b23f901dae377ca0a558e1e32aa56b31c4

This change unmasks netperf-2.7.0_p20210121 for amd64 and arm64, in
order to use the MIT licensed version of netperf.

Also, added netperf to the github actions sync list so that it gets
automatically synced with the upstream Gentoo.

Fixes: https://github.com/flatcar/Flatcar/issues/1459

Signed-off-by: Adrian Vladu <avladu@cloudbasesolutions.com>