Support the SDK on arm64
Catalyst 4 adds support for building with QEMU, so I initially leveraged
this, but it turned out to be very much slower than emulating the amd64
SDK on arm64, where an arm64 build could then be mostly run without
emulation. I have kept the code for the slower approach anyway since it
is small and may be useful to somebody.

There were several places where we assumed that amd64 was native and
arm64 required emulation via QEMU. The scripts are now more
architecture-agnostic, paving the way for riscv support later.

We no longer set QEMU_LD_PREFIX because it prevents the SDK itself from
being emulated. It also assumes there is only one non-native target,
which may not always be the case. bubblewrap does a better job of
running binaries under QEMU.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>