Publish stage 3 SDK tarball to use as a seed instead of stage4
We currently use a stage 4 tarball as a seed for building the SDK, but
stage 3 is smaller and less likely to cause issues.

This doesn't change the name of the tarball because supporting both
stage3 and stage4 at the same time just isn't feasible with the way the
digest check works.

The change is almost backwards compatible except that Catalyst will fail
to build older baselayout versions due to stage3 lacking git. A recent
change to the baselayout package sources it from a tarball instead of
git, and this has now made it into a release. This issue only affects
building the SDK, and it can be resolved by simply rebasing.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>