sys-apps/baselayout: Don't fetch non-live version from git
Use a regular tarball from GitHub. baselayout is a special case because
Catalyst initially installs it with --nodeps. We currently use stage 4
as a seed, which includes git, so this is fine. You might used a
cross-compiled stage 3 as a seed when porting the SDK to a new
architecture though, and this lacks git. We may also want to switch to
using stage 3 normally.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>