sys-devel/sysroot-wrappers: Drop in favour of setting CPP/CC/CXX
cross-boss has shown that setting CPP/CC/CXX is a highly effective
alternative to using wrappers. This approach is not susceptible to the
PATH or SYSROOT variables being lost. It works so well because Gentoo
uses a similar approach for amd64 multilib and failure to respect these
variables is treated as a serious QA violation.

On the downside, it can lead to the --sysroot argument leaking into
certain files and subsequently being applied to builds started from the
target environment. This can be worked around with a symlink, but it is
tidier to strip such flags after the build. cross-boss does this with a
pkg_preinst hook that carefully strips them from non-binary files. This
hook has been adapted here, but only for the dev profile because we
don't build the SDK with sysroot flags and we don't install build tools
to the production image. It has been tested against binary packages.

There has been talk in Gentoo circles of achieving this using gcc specs
instead, which would not suffer from the downside, but my initial
investigation has found that you cannot simply add --sysroot this way,
and it may not be viable at all.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>