sys-auth/realmd: Fix cross-compiling by pointing to krb5-config location
pkg-config is usually preferable, but this simpler approach works here
because krb5-config is a shell script and the headers and libraries are
in standard paths.

It was previously relying on a wrapper we were creating. It still works
without that, but only because krb5-config also exists on the host, and
we shouldn't assume that.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>