sys-block/open-iscsi: Migrate coreos-stable to portage-stable, 2.1.10
The old version 2.1.4 does not cross-compile without help from QEMU that
we cannot rely on going forwards. 2.1.10 is Meson-based and handles this
much better.

Rather than update the package in-place, migrate it to portage-stable
and cover the differences with a small patch and env script.

Upstream now carries the systemd files, so we do not need to add these.

/etc is now automatically moved to /usr/share/flatcar/etc, so we no
longer need any special handling for that here, but I have added a
compatibility symlink for iscsid.conf.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>