net-firewall/conntrack-tools: remove obsolete 1.4.5
This change removes an outdated conntrack-tools ebuild - version 1.4.5 -
from maintenance branch flatcar-4054. The branch uses 1.4.6-rc1 anyways
so the legacy ebuild can safely be deleted.

This ebuild uses EAPI 6 and will cause a transient error message during
build. It imports the linux-info eclass which was recently updated to
only support EAPI 7 and higher. This will cause emerge to print an
error message when considering the 1.4.5 ebuild. While that error
message has no impact on the build because 1.4.6-rc1 is more recent and
will hence be used, it can be misleading when investigating unrelated
issues (which it did to me, hence this patch).

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>