config/env/sys-fs/zfs-kmod: un-break zfs-kmod binpkg install
This change sets SKIP_KERNEL_BINPKG_ENV_RESET in the zfs-kmod env so
linux-info.eclass keeps kernel env variables. This resolves an issue
with installing zfs-kmod as a binpkg when the kernel was not yet
installed. In a pure binpkg install the zfs-kmod package might be
installed before the kernel, leading to the kmod's kernel version
detection to fail.

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>