btrfs hack: shrink the allocated size
We are using read-only btrfs compressed for the /usr partition,
and sometimes the free space shown by `btrfs fi usage` and the `df` IS diverging,
with `df` showing the free space from the `btrfs fi usage`'s `Device unallocated.

I have found a way to fix the allocation, by decreasing the size to more than the btrfs
filesystem allows, but curiosly the allocation is getting reset during the failed
attempt and the issue is fixed.

This is of course, a _HACK_, and can at any time produce unwanted behaviours.

Note that the issue cannot be reliably reproduced.

See Flatcar issue: https://github.com/flatcar/Flatcar/issues/1473

See issue upstream:
https://bugzilla.kernel.org/show_bug.cgi?id=219167

Signed-off-by: Adrian Vladu <avladu@cloudbasesolutions.com>