sys-boot/grub: Move to portage-stable, bump to 2.12, add Red Hat patches
We can now use Gentoo's upstream ebuild, save for a few small overrides
in a separate env file.

This bumps GRUB from 2.06 to 2.12, but with Red Hat's large patch set
applied, which is needed for Secure Boot. The existing two Flatcar
patches have been rebased. Gentoo's patches are discarded because they
conflict and are not relevant to Flatcar.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>