sys-boot/grub: Adjust version string to include flatcar1 suffix
We will bump this every time we or Gentoo change patches that modify
parts of GRUB that are installed to the boot partition. We will reset
the version back to 1 when the upstream GRUB version changes.

Without this, we are bound by Gentoo's PVR string, which we cannot
change when we need to make changes to our own patches. The shim review
wants to know the full version number of our GRUB build, and it would
look bad to make such changes without changing the version.

This suffix is also applied to the Flatcar entry in the SBAT, which is
especially important for the shim review.

The published binary package will still be labelled with the Gentoo PVR,
but that seems less important given that end users cannot update
individual packages on Flatcar installations.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>